--- notebooks/tools.py
@@ -1,432 +1,0 @@
-import pandas as pd
-import numpy as np
-import mlflow
-import matplotlib.pyplot as plt
-from sklearn.metrics import classification_report
-from sklearn.model_selection import KFold, StratifiedKFold
-from sklearn.model_selection import KFold
-from yellowbrick.classifier import ConfusionMatrix
-import pickle
-import gc
-import re
-import time
-from sklearn.metrics import *
-
-
-path = "../data_origin/"
-
-
-# Fonction d'encodage :
-def one_hot_encoder(df, nan_as_category=True):
-    original_columns = list(df.columns)
-    categorical_columns = [col for col in df.columns if df[col].dtype == "object"]
-    df = pd.get_dummies(df, columns=categorical_columns, dummy_na=nan_as_category)
-    new_columns = [c for c in df.columns if c not in original_columns]
-    return df, new_columns
-
-
-# Importation et transformation de application_train et application_test :
-def application_train_test(num_rows=None, nan_as_category=False):
-    # Read data and merge
-    df = pd.read_csv(path + "application_train.csv", nrows=num_rows)
-    test_df = pd.read_csv(path + "application_test.csv", nrows=num_rows)
-    print("Train set: {}, Test set: {}".format(len(df), len(test_df)))
-    df = df.append(test_df).reset_index()
-
-    # Remove "CODE_GENDER"
-    # df.drop(columns=["CODE_GENDER"], inplace=True)
-    # test_df.drop(columns=["CODE_GENDER"], inplace=True)
-
-    # Categorical features with Binary encode (0 or 1; two categories)
-    for bin_feature in ["FLAG_OWN_CAR", "FLAG_OWN_REALTY"]:
-        df[bin_feature], uniques = pd.factorize(df[bin_feature])
-
-    # Categorical features with One-Hot encode
-    df, cat_cols = one_hot_encoder(df, nan_as_category)
-
-    # NaN values for DAYS_EMPLOYED: 365.243 -> nan
-    df["DAYS_EMPLOYED"].replace(365243, np.nan, inplace=True)
-
-    # Some simple new features (percentages)
-    df["DAYS_EMPLOYED_PERC"] = df["DAYS_EMPLOYED"] / df["DAYS_BIRTH"]
-    df["INCOME_CREDIT_PERC"] = df["AMT_INCOME_TOTAL"] / df["AMT_CREDIT"]
-    df["INCOME_PER_PERSON"] = df["AMT_INCOME_TOTAL"] / df["CNT_FAM_MEMBERS"]
-    df["ANNUITY_INCOME_PERC"] = df["AMT_ANNUITY"] / df["AMT_INCOME_TOTAL"]
-    df["PAYMENT_RATE"] = df["AMT_ANNUITY"] / df["AMT_CREDIT"]
-    del test_df
-    gc.collect()
-    return df
-
-
-# Importation et transformation de bureau et bureau_balance :
-def bureau_and_balance(num_rows=None, nan_as_category=True):
-    bureau = pd.read_csv(path + "bureau.csv", nrows=num_rows)
-    bb = pd.read_csv(path + "bureau_balance.csv", nrows=num_rows)
-    bb, bb_cat = one_hot_encoder(bb, nan_as_category)
-    bureau, bureau_cat = one_hot_encoder(bureau, nan_as_category)
-
-    # Bureau balance: Perform aggregations and merge with bureau.csv
-    bb_aggregations = {"MONTHS_BALANCE": ["min", "max", "size"]}
-    for col in bb_cat:
-        bb_aggregations[col] = ["mean"]
-    bb_agg = bb.groupby("SK_ID_BUREAU").agg(bb_aggregations)
-    bb_agg.columns = pd.Index(
-        [e[0] + "_" + e[1].upper() for e in bb_agg.columns.tolist()]
-    )
-    bureau = bureau.join(bb_agg, how="left", on="SK_ID_BUREAU")
-    bureau.drop(["SK_ID_BUREAU"], axis=1, inplace=True)
-    del bb, bb_agg
-    gc.collect()
-
-    # Bureau and bureau_balance numeric features
-    num_aggregations = {
-        "DAYS_CREDIT": ["min", "max", "mean", "var"],
-        "DAYS_CREDIT_ENDDATE": ["min", "max", "mean"],
-        "DAYS_CREDIT_UPDATE": ["mean"],
-        "CREDIT_DAY_OVERDUE": ["max", "mean"],
-        "AMT_CREDIT_MAX_OVERDUE": ["mean"],
-        "AMT_CREDIT_SUM": ["max", "mean", "sum"],
-        "AMT_CREDIT_SUM_DEBT": ["max", "mean", "sum"],
-        "AMT_CREDIT_SUM_OVERDUE": ["mean"],
-        "AMT_CREDIT_SUM_LIMIT": ["mean", "sum"],
-        "AMT_ANNUITY": ["max", "mean"],
-        "CNT_CREDIT_PROLONG": ["sum"],
-        "MONTHS_BALANCE_MIN": ["min"],
-        "MONTHS_BALANCE_MAX": ["max"],
-        "MONTHS_BALANCE_SIZE": ["mean", "sum"],
-    }
-    # Bureau and bureau_balance categorical features
-    cat_aggregations = {}
-    for cat in bureau_cat:
-        cat_aggregations[cat] = ["mean"]
-    for cat in bb_cat:
-        cat_aggregations[cat + "_MEAN"] = ["mean"]
-
-    bureau_agg = bureau.groupby("SK_ID_CURR").agg(
-        {**num_aggregations, **cat_aggregations}
-    )
-    bureau_agg.columns = pd.Index(
-        ["BURO_" + e[0] + "_" + e[1].upper() for e in bureau_agg.columns.tolist()]
-    )
-
-    # Bureau: Active credits - using only numerical aggregations
-    active = bureau[bureau["CREDIT_ACTIVE_Active"] == 1]
-    active_agg = active.groupby("SK_ID_CURR").agg(num_aggregations)
-    active_agg.columns = pd.Index(
-        ["ACTIVE_" + e[0] + "_" + e[1].upper() for e in active_agg.columns.tolist()]
-    )
-    bureau_agg = bureau_agg.join(active_agg, how="left", on="SK_ID_CURR")
-    del active, active_agg
-    gc.collect()
-
-    # Bureau: Closed credits - using only numerical aggregations
-    closed = bureau[bureau["CREDIT_ACTIVE_Closed"] == 1]
-    closed_agg = closed.groupby("SK_ID_CURR").agg(num_aggregations)
-    closed_agg.columns = pd.Index(
-        ["CLOSED_" + e[0] + "_" + e[1].upper() for e in closed_agg.columns.tolist()]
-    )
-    bureau_agg = bureau_agg.join(closed_agg, how="left", on="SK_ID_CURR")
-    del closed, closed_agg, bureau
-    gc.collect()
-    return bureau_agg
-
-
-# Importation et transformation de bureau et bureau_balance :
-def previous_applications(num_rows=None, nan_as_category=True):
-    prev = pd.read_csv(path + "previous_application.csv", nrows=num_rows)
-    prev, cat_cols = one_hot_encoder(prev, nan_as_category=True)
-
-    # Days 365.243 values -> nan
-    prev["DAYS_FIRST_DRAWING"].replace(365243, np.nan, inplace=True)
-    prev["DAYS_FIRST_DUE"].replace(365243, np.nan, inplace=True)
-    prev["DAYS_LAST_DUE_1ST_VERSION"].replace(365243, np.nan, inplace=True)
-    prev["DAYS_LAST_DUE"].replace(365243, np.nan, inplace=True)
-    prev["DAYS_TERMINATION"].replace(365243, np.nan, inplace=True)
-
-    # Add feature: value ask / value received percentage
-    prev["APP_CREDIT_PERC"] = prev["AMT_APPLICATION"] / prev["AMT_CREDIT"]
-
-    # Previous applications numeric features
-    num_aggregations = {
-        "AMT_ANNUITY": ["min", "max", "mean"],
-        "AMT_APPLICATION": ["min", "max", "mean"],
-        "AMT_CREDIT": ["min", "max", "mean"],
-        "APP_CREDIT_PERC": ["min", "max", "mean", "var"],
-        "AMT_DOWN_PAYMENT": ["min", "max", "mean"],
-        "AMT_GOODS_PRICE": ["min", "max", "mean"],
-        "HOUR_APPR_PROCESS_START": ["min", "max", "mean"],
-        "RATE_DOWN_PAYMENT": ["min", "max", "mean"],
-        "DAYS_DECISION": ["min", "max", "mean"],
-        "CNT_PAYMENT": ["mean", "sum"],
-    }
-    # Previous applications categorical features
-    cat_aggregations = {}
-    for cat in cat_cols:
-        cat_aggregations[cat] = ["mean"]
-
-    prev_agg = prev.groupby("SK_ID_CURR").agg({**num_aggregations, **cat_aggregations})
-    prev_agg.columns = pd.Index(
-        ["PREV_" + e[0] + "_" + e[1].upper() for e in prev_agg.columns.tolist()]
-    )
-
-    # Previous Applications: Approved Applications - only numerical features
-    approved = prev[prev["NAME_CONTRACT_STATUS_Approved"] == 1]
-    approved_agg = approved.groupby("SK_ID_CURR").agg(num_aggregations)
-    approved_agg.columns = pd.Index(
-        ["APPROVED_" + e[0] + "_" + e[1].upper() for e in approved_agg.columns.tolist()]
-    )
-    prev_agg = prev_agg.join(approved_agg, how="left", on="SK_ID_CURR")
-
-    # Previous Applications: Refused Applications - only numerical features
-    refused = prev[prev["NAME_CONTRACT_STATUS_Refused"] == 1]
-    refused_agg = refused.groupby("SK_ID_CURR").agg(num_aggregations)
-    refused_agg.columns = pd.Index(
-        ["REFUSED_" + e[0] + "_" + e[1].upper() for e in refused_agg.columns.tolist()]
-    )
-    prev_agg = prev_agg.join(refused_agg, how="left", on="SK_ID_CURR")
-    del refused, refused_agg, approved, approved_agg, prev
-    gc.collect()
-    return prev_agg
-
-
-def pos_cash(num_rows=None, nan_as_category=True):
-    pos = pd.read_csv(path + "POS_CASH_balance.csv", nrows=num_rows)
-    pos, cat_cols = one_hot_encoder(pos, nan_as_category=True)
-
-    # Features
-    aggregations = {
-        "MONTHS_BALANCE": ["max", "mean", "size"],
-        "SK_DPD": ["max", "mean"],
-        "SK_DPD_DEF": ["max", "mean"],
-    }
-    for cat in cat_cols:
-        aggregations[cat] = ["mean"]
-
-    pos_agg = pos.groupby("SK_ID_CURR").agg(aggregations)
-    pos_agg.columns = pd.Index(
-        ["POS_" + e[0] + "_" + e[1].upper() for e in pos_agg.columns.tolist()]
-    )
-
-    # Count pos cash accounts
-    pos_agg["POS_COUNT"] = pos.groupby("SK_ID_CURR").size()
-    del pos
-    gc.collect()
-    return pos_agg
-
-
-def installments_payments(num_rows=None, nan_as_category=True):
-    ins = pd.read_csv(path + "installments_payments.csv", nrows=num_rows)
-    ins, cat_cols = one_hot_encoder(ins, nan_as_category=True)
-
-    # Percentage and difference paid in each installment (amount paid and installment value)
-    ins["PAYMENT_PERC"] = ins["AMT_PAYMENT"] / ins["AMT_INSTALMENT"]
-    ins["PAYMENT_DIFF"] = ins["AMT_INSTALMENT"] - ins["AMT_PAYMENT"]
-
-    # Days past due and days before due (no negative values)
-    ins["DPD"] = ins["DAYS_ENTRY_PAYMENT"] - ins["DAYS_INSTALMENT"]
-    ins["DBD"] = ins["DAYS_INSTALMENT"] - ins["DAYS_ENTRY_PAYMENT"]
-    ins["DPD"] = ins["DPD"].apply(lambda x: x if x > 0 else 0)
-    ins["DBD"] = ins["DBD"].apply(lambda x: x if x > 0 else 0)
-
-    # Features: Perform aggregations
-    aggregations = {
-        "NUM_INSTALMENT_VERSION": ["nunique"],
-        "DPD": ["max", "mean", "sum"],
-        "DBD": ["max", "mean", "sum"],
-        "PAYMENT_PERC": ["max", "mean", "sum", "var"],
-        "PAYMENT_DIFF": ["max", "mean", "sum", "var"],
-        "AMT_INSTALMENT": ["max", "mean", "sum"],
-        "AMT_PAYMENT": ["min", "max", "mean", "sum"],
-        "DAYS_ENTRY_PAYMENT": ["max", "mean", "sum"],
-    }
-    for cat in cat_cols:
-        aggregations[cat] = ["mean"]
-    ins_agg = ins.groupby("SK_ID_CURR").agg(aggregations)
-    ins_agg.columns = pd.Index(
-        ["INSTAL_" + e[0] + "_" + e[1].upper() for e in ins_agg.columns.tolist()]
-    )
-
-    # Count installments accounts
-    ins_agg["INSTAL_COUNT"] = ins.groupby("SK_ID_CURR").size()
-    del ins
-    gc.collect()
-    return ins_agg
-
-
-def credit_card_balance(num_rows=None, nan_as_category=True):
-    cc = pd.read_csv(path + "credit_card_balance.csv", nrows=num_rows)
-    cc, cat_cols = one_hot_encoder(cc, nan_as_category=True)
-
-    # General aggregations
-    cc.drop(["SK_ID_PREV"], axis=1, inplace=True)
-    cc_agg = cc.groupby("SK_ID_CURR").agg(["min", "max", "mean", "sum", "var"])
-    cc_agg.columns = pd.Index(
-        ["CC_" + e[0] + "_" + e[1].upper() for e in cc_agg.columns.tolist()]
-    )
-
-    # Count credit card lines
-    cc_agg["CC_COUNT"] = cc.groupby("SK_ID_CURR").size()
-    del cc
-    gc.collect()
-    return cc_agg
-
-
-def main(debug=False):
-    num_rows = 10000 if debug else None
-
-    df = application_train_test(num_rows)
-    bureau = bureau_and_balance(num_rows)
-    print("Bureau df shape:", bureau.shape)
-    df = df.join(bureau, how="left", on="SK_ID_CURR")
-    del bureau
-    gc.collect()
-
-    prev = previous_applications(num_rows)
-    print("Previous applications df shape:", prev.shape)
-    df = df.join(prev, how="left", on="SK_ID_CURR")
-    del prev
-    gc.collect()
-
-    pos = pos_cash(num_rows)
-    print("Pos-cash balance df shape:", pos.shape)
-    df = df.join(pos, how="left", on="SK_ID_CURR")
-    del pos
-    gc.collect()
-
-    ins = installments_payments(num_rows)
-    print("Installments payments df shape:", ins.shape)
-    df = df.join(ins, how="left", on="SK_ID_CURR")
-    del ins
-    gc.collect()
-
-    cc = credit_card_balance(num_rows)
-    print("Credit card balance df shape:", cc.shape)
-    df = df.join(cc, how="left", on="SK_ID_CURR")
-    del cc
-    gc.collect()
-
-    df = df.rename(columns=lambda x: re.sub("[^A-Za-z0-9_]+", "", x))
-
-    # df.to_csv("data.csv")
-
-    return df
-
-
-def cross_val_split(X, y, num_folds=10, stratified=False, debug=False):
-    """This function implement a cross validation.It takes as input a some features(X),a target variable 
-    and integer as the number of folds. It return the trainings and testing sets after the cross validation."""
-    # Cross validation model
-    if stratified:
-        folds = StratifiedKFold(n_splits=num_folds, shuffle=True, random_state=5)
-    else:
-        folds = KFold(n_splits=num_folds, shuffle=True, random_state=5)
-    for n_fold, (train_idx, valid_idx) in enumerate(folds.split(X, y)):
-        train_x, train_y = X.iloc[train_idx], y.iloc[train_idx]
-        valid_x, valid_y = X.iloc[valid_idx], y.iloc[valid_idx]
-    for key, value in {
-        "X_train": train_x,
-        "X_test": valid_x,
-        "y_train": train_y,
-        "y_test": valid_y,
-    }.items():
-        print(f"{key} shape : {value.shape}")
-    return train_x, valid_x, train_y, valid_y
-
-
-# Cette fonction a pour objectif d'afficher un aperçu et une description d'un dataframe ainsi que le nbre de missing values qu'il contient
-def describe_data(df, figsize=(6, 4)):
-    # print('*'*35,'Data infos','*'*35)
-
-    # Check nombre de colonnes
-    print("Nombre de colonnes : ", df.shape[1], "\n")
-
-    # Check nombre de lignes
-    print("Nombre de lignes : ", df.shape[0], "\n")
-
-    # Analyse des valeurs manquantes
-    plt.figure(figsize=(9, 6))
-    # print('*'*34,"Valeurs manquantes",'*'*34)
-    all_df = df.isnull().sum().sum(), df.notnull().sum().sum()
-    plt.pie(
-        all_df,
-        autopct="%1.1f%%",
-        shadow=False,
-        startangle=90,
-        labels=["Missing values", "Not missing values"],
-        explode=(0, 0.02),
-        colors=["lightblue", "steelblue"],
-        pctdistance=0.4,
-        labeldistance=1.1,
-    )
-    circle = plt.Circle((0, 0), 0.65, color="white")
-    p = plt.gcf()
-    p.gca().add_artist(circle)
-    plt.show()
-
-    print("Nombre total de valeurs manquantes : ", df.isna().sum().sum(), "\n")
-
-    # run with normal data
-
-
-
-        
-        
-
-def evaluate_model(model, x, y, x_test, y_test, model_name, balancing_method):
-    
-    #Entrainement
-    start = time.time()
-    #model = gs.fit(x,y)
-    end = time.time()-start
-
-    if (model_name != 'Baseline'):
-        df_results = pd.DataFrame.from_dict(model.cv_results_)
-
-    #Training Performance
-    if (model_name == 'Baseline'):
-        #y_pred = model.predict(x)
-        y_proba = model.predict_proba(x)
-
-        auc_train = round(roc_auc_score(y, y_proba[:,1]),3) 
-        #f2_train = round(fbeta_score(y, y_pred, beta=2), 3)
-    else:
-        auc_train = round(model.best_score_,3) 
-        #f2_train = round(np.mean(df_results[df_results.rank_test_F2 == 1]['mean_test_F2']),3)
-
-    #Testing Performance
-    #y_pred = model.predict(x_test)
-    y_proba = model.predict_proba(x_test)
-    auc_test = round(roc_auc_score(y_test, y_proba[:,1]),3) 
-    #f2_test = round(fbeta_score(y_test, y_pred, beta=2), 3)
-
-    row = [model_name, 
-            balancing_method,
-            auc_train, 
-            auc_test,
-            #f2_train,
-            #f2_test,
-            end]
-
-    return row
-
-#------------------------------------------
-
-
-def run_experiment(experiment_name, name, model, X_train, X_test, y_train, y_test, model_name, balancing_method):
-    mlflow.set_experiment(experiment_name)
-    mlflow.sklearn.autolog()
-    mlflow.lightgbm.autolog()
-    with mlflow.start_run(run_name=name):
-        mlflow.set_tag("delevoper", "Alassane")
-        my_model = model
-        my_model.fit(X_train, y_train)
-        y_pred = my_model.predict(X_test)
-        print("*" * 25, f"Test scores - {name}", "*" * 25)
-        print(classification_report(y_test, y_pred))
-        cm = ConfusionMatrix(my_model, classes=y_train.value_counts().index)
-        cm.score(X_test, y_test)
-        cm.show()
-        metrics = evaluate_model(my_model, X_train, y_train, X_test, y_test, model_name, balancing_method)
-        return metrics, my_model
-    
-
