+++ pyproject.toml
@@ -1,10 +1,10 @@
 [tool.poetry]
-name = "loan-credit-scoring"
+name = "loan-scoring"
 version = "0.1.0"
 description = ""
 authors = ["Alhassanne <alhassanne.ahmed@servier.com>"]
 readme = "README.md"
-packages = [{include = "loan_credit_scoring"}]
+packages = [{include = "loan_scoring"}]
 
 [tool.poetry.dependencies]
 python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
@@ -25,8 +25,9 @@ hyperopt = "^0.2.7"
 shap = "^0.41.0"
 numba = "^0.56.4"
 boltons = "^23.0.0"
-evidently = "^0.3.0"
 jupyter-contrib-nbextensions = "^0.7.0"
+scipy = "^1.10.1"
+cinnamon = "^0.2.1"
 
 
 [build-system]
